Added bad case for msg sig
